made tests work with abtio and memory logs

+++ tests/run-test.sh
@@ -1,14 +1,23 @@
 #!/bin/bash
 
 SCENARIO=$1
-STORAGE=`mktemp -d test-log-XXXXXXXX`
+BACKEND=$2
 
-timeout 600 ./mraft-test na+sm -n 3 -f $1 -p $STORAGE
+if ["$BACKEND" = "abt-io"]; then
+    STORAGE=`mktemp -d test-log-XXXXXXXX`
+    STORAGE="-p $STORAGE"
+else
+    STORAGE=""
+fi
+
+timeout 600 ./mraft-test na+sm -n 3 -f $SCENARIO $STORAGE -l $BACKEND
 RET=$?
 
 if [ $RET -eq 0 ]
 then
-    rm -rf $STORAGE
+    if [[ -n $STORAGE ]]; then
+        rm -rf $STORAGE
+    fi
 fi
 
 exit $RET
